freeze migrations into v1.23.0_2.ts, bump to 1.23.0:3
